Added serverless support for 'python3.6' runtime in AWS
To use this, source the serverless-bash.sh script in .bash_profile,
which will provide the 'sls' command/function in Bash. That function
will run the correct Docker container based on the runtime configured in
the serverless.yml file (sls command must be run from the root directory
of the serverless project).